requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
fpdf==1.7.2
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
watchdog==6.0.0
